+++ compiler/qsc_passes/src/capabilitiesck/tests_adaptive_plus_integers.rs
@@ -23,7 +23,7 @@ fn check_profile(source: &str, expect: &Expect) {
     check(
         source,
         expect,
-        RuntimeCapabilityFlags::ForwardBranching | RuntimeCapabilityFlags::IntegerComputations,
+        RuntimeCapabilityFlags::Adaptive | RuntimeCapabilityFlags::IntegerComputations,
     );
 }
