+++ library/std/convert.qs
@@ -27,7 +27,7 @@ namespace Microsoft.Quantum.Convert {
     ///
     /// # Output
     /// A `Bool` representing the `input`.
-    @Config(ForwardBranching)
+    @Config(Adaptive)
     function ResultAsBool(input : Result) : Bool {
         input == One
     }
@@ -42,7 +42,7 @@ namespace Microsoft.Quantum.Convert {
     ///
     /// # Output
     /// A `Result` representing the `input`.
-    @Config(ForwardBranching)
+    @Config(Adaptive)
     function BoolAsResult(input : Bool) : Result {
         if input { One } else { Zero }
     }
@@ -168,8 +168,7 @@ namespace Microsoft.Quantum.Convert {
     /// // The following returns 1
     /// let int1 = ResultArrayAsInt([One,Zero])
     /// ```
-    @Config(ForwardBranching)
-    @Config(IntegerComputations)
+    @Config(Adaptive)
     function ResultArrayAsInt(results : Result[]) : Int {
         let nBits = Length(results);
         Fact(nBits < 64, $"`Length(bits)` must be less than 64, but was {nBits}.");
@@ -194,7 +193,7 @@ namespace Microsoft.Quantum.Convert {
     ///
     /// # Output
     /// A `Bool[]` representing the `input`.
-    @Config(ForwardBranching)
+    @Config(Adaptive)
     function ResultArrayAsBoolArray(input : Result[]) : Bool[] {
         mutable output = [];
         for r in input {
@@ -214,7 +213,7 @@ namespace Microsoft.Quantum.Convert {
     ///
     /// # Output
     /// A `Result[]` representing the `input`.
-    @Config(ForwardBranching)
+    @Config(Adaptive)
     function BoolArrayAsResultArray(input : Bool[]) : Result[] {
         mutable output = [];
         for b in input {
