+++ library/std/unstable_arithmetic_internal.qs
@@ -243,13 +243,10 @@ namespace Microsoft.Quantum.Unstable.Arithmetic {
     ///   Phys. Rev. A 87, 022328, 2013
     ///   [arXiv:1212.5069](https://arxiv.org/abs/1212.5069)
     ///   doi:10.1103/PhysRevA.87.022328
-    @Config(ForwardBranching)
+    @Config(Adaptive)
     internal operation ApplyAndAssuming0Target(control1 : Qubit, control2 : Qubit, target : Qubit) : Unit is Adj {
         // NOTE: Eventually this operation will be public and intrinsic.
         body (...) {
-            if not CheckZero(target) {
-                fail "ApplyAndAssuming0Target expects `target` to be in |0> state.";
-            }
             CCNOT(control1, control2, target);
         }
         adjoint (...) {
