+++ library/std/unstable_table_lookup.qs
@@ -43,7 +43,7 @@ namespace Microsoft.Quantum.Unstable.TableLookup {
     ///     "Windowed arithmetic"
     /// [3] [arXiv:2211.01133](https://arxiv.org/abs/2211.01133)
     ///     "Space-time optimized table lookup"
-    @Config(ForwardBranching)
+    @Config(Adaptive)
     operation Select(
         data : Bool[][],
         address : Qubit[],
@@ -97,7 +97,6 @@ namespace Microsoft.Quantum.Unstable.TableLookup {
         }
     }
 
-    @Config(ForwardBranching)
     internal operation SinglyControlledSelect(
         ctl : Qubit,
         data : Bool[][],
@@ -166,7 +165,7 @@ namespace Microsoft.Quantum.Unstable.TableLookup {
     /// # References
     /// - [arXiv:1905.07682](https://arxiv.org/abs/1905.07682)
     ///   "Windowed arithmetic"
-    @Config(ForwardBranching)
+    @Config(Adaptive)
     internal operation Unlookup(
         lookup : (Bool[][], Qubit[], Qubit[]) => Unit,
         data : Bool[][],
